+++ cannon/src/main/java/cannongame/ICalculateYValue.java
@@ -1,5 +1,5 @@
 package cannongame;
 
 public interface ICalculateYValue {
-    public double YValue(double degree, int velocity);
+    public double YValue(double degree, double velocity);
 }
